Refactor(Telescope): Move Telescope configuration into separate files
This change moves Telescope plugin configuration and Telescope-related
LSP keybindings into dedicated files. Separating plugin configuration
simplifies future refactoring (e.g. replacing Telescope with Snacks).
